cache: move this test to storage_test.go
